better fastapi routing
